Implemented UnitLevel figure report - added into `mapshell`
