+++ pipeline/plot/__init__.py
@@ -1,12 +1,14 @@
 
 from .util import show_source
 
-from .unit_psth import unit_psth
-from .unit_psth import unit_psth_ll
+# from .unit_psth import unit_psth
+# from .unit_psth import unit_psth_ll
+#
+# from .group_psth import group_psth
+# from .group_psth import group_psth_ll
+#
+# __all__ = [show_source,
+#            unit_psth, unit_psth_ll,
+#            group_psth, group_psth_ll]
 
-from .group_psth import group_psth
-from .group_psth import group_psth_ll
-
-__all__ = [show_source,
-           unit_psth, unit_psth_ll,
-           group_psth, group_psth_ll]
+__all__ = [show_source]
