+++ pipeline/report.py
@@ -5,9 +5,10 @@ import pathlib
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
+from matplotlib.gridspec import GridSpec
 
 from pipeline import experiment, ephys, psth, tracking
-from pipeline.plot import behavior_plot, unit_characteristic_plot
+from pipeline.plot import behavior_plot, unit_characteristic_plot, unit_psth
 from pipeline import get_schema_name
 
 schema = dj.schema(get_schema_name('report'))
@@ -29,7 +30,7 @@ class SessionLevelReport(dj.Computed):
     key_source = experiment.Session & experiment.BehaviorTrial & tracking.Tracking
 
     def make(self, key):
-        sess_dir = store_directory / str(key['subject_id']) / str(key['session'])
+        sess_dir = store_directory / f'subject_{key["subject_id"]}' / f'session_{key["session"]}'
         sess_dir.mkdir(parents=True, exist_ok=True)
 
         # ---- behavior_performance ----
@@ -75,12 +76,20 @@ class ProbeLevelReport(dj.Computed):
     ---
     clustering_quality: filepath@report_store
     unit_characteristic: filepath@report_store
+    group_psth: filepath@report_store
     """
 
-    key_source = ephys.ProbeInsertion & ephys.UnitStat & psth.UnitSelectivity
+    @property
+    def key_source(self):
+        # Only process ProbeInsertion with UnitSelectivity computation fully completed
+        ks = (ephys.ProbeInsertion & ephys.UnitStat).proj()
+        unit = ks.aggr(ephys.Unit & 'unit_quality != "all"', unit_count='count(*)')
+        sel_unit = ks.aggr(psth.UnitSelectivity, sel_unit_count='count(*)')
+        return unit * sel_unit & 'unit_count = sel_unit_count'
 
     def make(self, key):
-        sess_dir = store_directory / str(key['subject_id']) / str(key['session']) / str(key['insertion_number'])
+        sess_dir = (store_directory / f'subject_{key["subject_id"]}' / f'session_{key["session"]}'
+                    / f'probe_{key["insertion_number"]}')
         sess_dir.mkdir(parents=True, exist_ok=True)
 
         probe_insertion = ephys.ProbeInsertion & key
@@ -104,16 +113,23 @@ class ProbeLevelReport(dj.Computed):
         unit_characteristic_plot.plot_unit_selectivity(probe_insertion, axs=axs[:3])
         unit_characteristic_plot.plot_unit_bilateral_photostim_effect(probe_insertion, axs=axs[-1])
 
+        # ---- unit_characteristic ----
+        fig3 = unit_characteristic_plot.plot_stacked_contra_ipsi_psth(ephys.Unit & key)
+
         # ---- Save fig and insert ----
         fig1_fp = sess_dir / 'clustering_quality.png'
         fig2_fp = sess_dir / 'unit_characteristic.png'
+        fig3_fp = sess_dir / 'group_psth.png'
 
         fig1.savefig(fig1_fp)
         print(f'Generated {fig1_fp}')
         fig2.savefig(fig2_fp)
         print(f'Generated {fig2_fp}')
+        fig3.savefig(fig3_fp)
+        print(f'Generated {fig3_fp}')
         self.insert1(dict(key, clustering_quality=fig1_fp.as_posix(),
-                          unit_characteristic=fig2_fp.as_posix()))
+                          unit_characteristic=fig2_fp.as_posix(),
+                          group_psth=fig3_fp.as_posix()))
 
 
 @schema
@@ -121,13 +137,42 @@ class UnitLevelReport(dj.Computed):
     definition = """
     -> ephys.ProbeInsertion
     ---
-    clustering_quality: filepath@report_store
-    unit_characteristic: filepath@report_store
+    unit_psth: filepath@report_store
+    unit_behavior: filepath@report_store
     """
 
     key_source = ephys.Unit & psth.UnitSelectivity
 
     def make(self, key):
-        pass
+        sess_dir = (store_directory / f'subject_{key["subject_id"]}' / f'session_{key["session"]}'
+                    / f'probe_{key["insertion_number"]}' / f'unit_{key["unit"]}')
+        sess_dir.mkdir(parents=True, exist_ok=True)
+
+        fig1 = unit_psth.plot_unit_psth(key)
+
+        fig2 = plt.figure(figsize = (16, 16))
+        gs = GridSpec(4, 2)
+
+        behavior_plot.plot_tracking(experiment.Session & key, key, tracking_feature='jaw_x', xlim=(-0.5, 1),
+                                    trial_offset=0, trial_limit=30, axs=np.array([fig2.add_subplot(gs[:3, col])
+                                                                                  for col in range(2)]))
+
+        behavior_plot.plot_unit_jaw_phase_dist(experiment.Session & key, key,
+                                               axs=np.array([fig2.add_subplot(gs[-1, col], polar=True)
+                                                             for col in range(2)]))
+
+        fig2.subplots_adjust(wspace=0.4)
+        fig2.subplots_adjust(hspace=0.6)
+
+        # ---- Save fig and insert ----
+        fig1_fp = sess_dir / 'unit_psth.png'
+        fig2_fp = sess_dir / 'unit_behavior.png'
+
+        fig1.savefig(fig1_fp)
+        print(f'Generated {fig1_fp}')
+        fig2.savefig(fig2_fp)
+        print(f'Generated {fig2_fp}')
+        self.insert1(dict(key, unit_psth=fig1_fp.as_posix(),
+                          unit_behavior=fig2_fp.as_posix()))
